fix: not work after change status many times
When you change the solved status, the problems may change to a little, then the container table may add rows after you change to another status. 
In cases like this, the Mutation Observer will observe the table change, not the row change. It means your observe `target` will be the table, not the columns. 
Besides, when `addedNodes` is not empty, it doesn't mean the table change, but also maybe one column include many changes.